Fix Arc weight conversion: preserve float values during load and init
CRITICAL BUG FIX: Arc.__init__() and Arc.from_dict() were converting
weight to int, truncating float values like 2.5 → 2.

Changes:
- Arc.__init__(): self.weight = float(weight) instead of int(weight)
- Arc.from_dict(): weight=float(data.get('weight', 1)) instead of int()

This caused inhibitor arcs with threshold 2.5 to behave as threshold 2,
allowing transitions to fire when they should be blocked.

Example: ATP inhibitor with weight=2.5 mM was truncated to 2,
so transitions fired at ATP=2.2 when they should wait until ATP<2.5.

+++ src/shypn/netobjs/arc.py
@@ -43,7 +43,7 @@ class Arc(PetriNetObject):
         # Connection (references to object instances)
         self.source = source
         self.target = target
-        self.weight = int(weight)
+        self.weight = float(weight)  # Float for precise thresholds (e.g., 2.5 mM)
         
         # Styling
         self.color = self.DEFAULT_COLOR
@@ -759,7 +759,7 @@ class Arc(PetriNetObject):
             target=target,
             id=arc_id,  # String ID
             name=arc_name,
-            weight=int(data.get("weight", 1))
+            weight=float(data.get("weight", 1))
         )
         
         # Restore optional properties
